requirements.txt
pyyaml
psycopg2-binary
svn
schedule
